+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Add `getListItemForItem` API to `CollectionComponentClass.js`.
 - Updated to @vcmap/core 5.1
 - Added viewshed plugin to demo application and external plugins
+- Added measurement plugin to demo application and external plugins
 
 ### Ui Behaviour ChangesF